# Use the database to check if a package was installed or not