@tgamblin 2021-05-28
@scheibelp @tldahlgren: can you take a look at this?  I am not remembering all the ins and outs of checking whether something is installed, but as you know it is complicated and we try to handle cases where, e.g., an install failed halfway through.

So I am not sure we want to do this, as we spend an awfully long time getting it right.

@haampie's issue may be resolved by #22845, wherein we will be switching to `json`, which is much faster.  So I am not sure we want to shake things up quite this much on the installed check.