@haampie 2021-05-28
Following git blame, it seems the origin of that line is https://github.com/spack/spack/pull/504 and it was left intact during the distributed builds feature. So we may also want to pick @tgamblin's brain from that time? 😅