@haampie 2021-05-31
I think we should hold off from merging this, since it's not functionally equivalent to previous behavior due to a missing warning.

Couple questions I have:

1. Is it intended behavior that `spack install abc def` for different specs `abc` and `def` with coinciding install prefix does not error?
2. Why do we handle `InconsistentInstallDirectoryError` and `SpecHashCollisionError` at best as warnings (at best cause it's conditional on  `--keep-prefix`)?